refactor(mcp): extract list_tools to handlers_tools.rs (lab-kvji.24.1.4)
list_tools body -> inherent list_tools_impl in handlers_tools.rs, including the
gateway meta-tool search/execute input-schema construction. ServerHandler keeps a
one-line delegator. CODE_EXECUTE_DESCRIPTION + CODE_MODE_MAX_CODE_BYTES widened to
pub(crate) in server.rs (single definition); handlers_tools imports the description.

handlers_tools.rs 212 LOC. server.rs 2382 -> 2203. 1423 tests pass.
